+++ libraries/ui_compose/src/main/java/androidx/media3/ui/compose/ContentFrame.kt
@@ -41,6 +41,7 @@ import androidx.media3.ui.compose.state.rememberPresentationState
  * @param surfaceType The type of surface to use for video playback. Can be either
  *   [SURFACE_TYPE_SURFACE_VIEW] or [SURFACE_TYPE_TEXTURE_VIEW].
  * @param contentScale The [ContentScale] strategy for the container.
+ * @param artwork Optional composable slot to render artwork for the current media item.
  * @param keepContentOnReset If `true`, the last rendered frame will remain visible when the player
  *   is reset. If `false`, the surface will be cleared.
  * @param overlay A composable drawn on top of the media content, but under the shutter.
@@ -54,6 +55,7 @@ fun ContentFrame(
   modifier: Modifier = Modifier,
   surfaceType: @SurfaceType Int = SURFACE_TYPE_SURFACE_VIEW,
   contentScale: ContentScale = ContentScale.Fit,
+  artwork: (@Composable (Player?) -> Unit)? = null,
   keepContentOnReset: Boolean = false,
   overlay: @Composable () -> Unit = {},
   shutter: @Composable () -> Unit = { Box(Modifier.fillMaxSize().background(Color.Black)) },
@@ -74,6 +76,10 @@ fun ContentFrame(
 
     if (presentationState.coverSurface) {
       shutter()
+    }
+
+    if (presentationState.showArtwork) {
+      artwork?.invoke(player)
     }
   }
 }